refactor(sort algorithms)

+++ lib/global-headers/CC-PRINT-01.h
@@ -1,8 +1,8 @@
-void printArray(int *Array, int lengthArray)
+void printArray(long int *Array, long int lengthArray)
 {
-    for (int i = 0; i < lengthArray; i++)
+    for (long int i = 0; i < lengthArray; i++)
     {
-        printf("| %.2d:  %.2d |", i, Array[i]);
+        printf("| %.2ld:  %.2ld |", i, Array[i]);
     }
     printf("\n");
 }
