+++ lib/sort-headers/CC-MERGE-02.h
@@ -1,12 +1,12 @@
 #include "CC-MERGE-03.h"
 
-void _mergeSort(int *Integers, int beggininterval, int endinterval, int *comparations, int *copies)
+void _mergeSort(long int *Integers, long int beggininterval, long int endinterval, long int *comparations, long int *copies)
 {
     *comparations = *comparations + 1;
 
     if (beggininterval < endinterval)
     {
-        int middleinterval = (beggininterval + endinterval) / 2;
+        long int middleinterval = (beggininterval + endinterval) / 2;
 
         _mergeSort(Integers, beggininterval, middleinterval, comparations, copies);
         _mergeSort(Integers, middleinterval + 1, endinterval, comparations, copies);
